+++ packages/tools/pluggable-widgets-tools/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+- We fixed the typing generation for actions that are referenced by the `onChange` of an attribute.
+
 ### Added
 - We added support for dark icons (`widgetName`.icon.dark.png and `widgetName`.tile.dark.png).